// src/types/dataPemetaanKebun.ts
export interface DataPemetaanKebun {
    idPemetaanKebun: string;
    nikSurveyor: string;
    namaSurveyor: string;
    longitude: string;
    latitude: string;
    statusKawasan: string;
    luasKebun: string;
    nipVerifikator: string;
    namaVerifikator: string;
    statusVerifikator: string;
    pesanVerifikator: string;
    waktuPemetaanKebun: string;
    waktuVerifikator: string;
    waktuUpdatePemetaanKebun: string;
    waktuUpdateVerifikator: string;
    IdTransaksiBlockchain: string;
}